[#428] PC/RS/VD Add tests for readDocsByCollection if collection does not exist
