+++ app/src/hooks/useProjectsFromFirestore.js
@@ -1,0 +1,25 @@
+import { CollectionNames } from "../datastore";
+
+const projectDb = async (datastore) => {
+  const extractProjectProperties = (project) => {
+    const { projectName } = project.data;
+    return projectName;
+  };
+
+  const projects = await datastore.readDocsFromCollection(
+    CollectionNames.PROJECT
+  );
+
+  console.log(projects);
+
+  if (!projects.length) {
+    return [{ entries: [] }];
+  }
+
+  const extractedProjects = projects.map((project) =>
+    extractProjectProperties(project)
+  );
+  return extractedProjects;
+};
+
+export default projectDb;
